Rename files: Drop 'helper' suffix
We are going to add more functionality to this file,
so the helper part of the name won't apply anymore.
